RemoteHarvesters now perform workPartJobs in remoteRoom before returning to homeRoom
